requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==11.0.0
Werkzeug==3.0.0
